requirements.txt
-find-links wheelhouse
wheel
azure==0.8.4
alembic==0.9.1
appdirs==1.4.3
click==6.7
Flask==0.12.1
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.1.9
Werkzeug==0.12.1
waitress==1.0.1
flask_sqlalchemy_session

numpy
flask_cors
Flask-Login==0.3.2
Flask-WTF==0.14.2
WTForms-Alchemy==0.16.6
WTForms-Components==0.10.3

flask-debugtoolbar==0.10.0
Flask-Mail==0.9.1